pyproject.toml
[project]
name = "trustyai"
version = "0.6.2"
description = "Python bindings to the TrustyAI explainability library."
authors = [{ name = "Rui Vieira", email = "rui@redhat.com" }]
license = { text = "Apache License Version 2.0" }
readme = "README.md"
requires-python = ">=3.10"

keywords = ["trustyai", "xai", "explainability", "ml"]

classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Java",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Java Libraries",
]

dependencies = [
    "Jpype1==1.5.0",
    "pyarrow==20.0.0",
    "matplotlib~=3.10.3",
    "pandas~=1.5.3",
    "numpy~=1.26.4",
    "jupyter-bokeh~=4.0.5",
]

[project.optional-dependencies]
dev = [
    "JPype1==1.5.0",
    "black~=22.12.0",
    "click==8.0.4",
    "joblib~=1.2.0",
    "jupyterlab~=4.4.4",
    "numpydoc==1.5.0",
    "pylint==3.2.0",
    "pytest~=7.2.1",
    "pytest-benchmark==4.0.0",
    "pytest-forked~=1.6.0",
    "scikit-learn~=1.7.0",
    "setuptools",
    "twine==3.4.2",
    "wheel~=0.38.4",
    "xgboost~=3.0.2",
]
extras = ["aix360[default,tsice,tslime,tssaliency]==0.3.0"]

detoxify = [
    "transformers~=4.36.2",
    "datasets",
    "scipy~=1.12.0",
    "torch~=2.2.1",
    "iter-tools",
    "evaluate",
    "trl",
]

api = ["kubernetes"]

[project.urls]
homepage = "https://github.com/trustyai-explainability/trustyai-explainability-python"
documentation = "https://trustyai-explainability-python.readthedocs.io/en/latest/"
repository = "https://github.com/trustyai-explainability/trustyai-explainability-python"

[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.pytest.ini_options]
log_cli = true
addopts = '-m="not block_plots"'
markers = [
    "block_plots: Test plots will block execution of subsequent tests until closed",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
